requirements.txt
python-telegram-bot==22.7
aiosqlite==0.21.0
